Build Product 6

+++ app/hometemplate/sanPham/chiTietSanPham/binhLuan/xoaBinhLuan.tsx
@@ -3,33 +3,37 @@ import getBearerToken from "@/app/helpers/LogicHelper/authHelper";
 import AppUser from "@/app/model/AppUser";
 import { url } from "@/app/server/backend";
 import axios from "axios";
-import { Button, Text, TouchableOpacity, View } from "react-native";
+import { DimensionValue, Text, TouchableOpacity, View } from "react-native";
 
-export default function XoaBinhLuan({userLogin, binhLuan, layCacBinhLuans} : {userLogin: AppUser, binhLuan: any, layCacBinhLuans: () => void}) {
-    const xoaBinhLuan = () => {
-        getBearerToken()
-        .then((bearerToken: any) => {
-            let urlXoaBinhLuan = url(`api/binhluan/${binhLuan.bL_Id}`);
+export default function XoaBinhLuan({userLogin, binhLuan, layCacBinhLuans, width} : {userLogin: AppUser | null, binhLuan: any, layCacBinhLuans: () => void, width: DimensionValue | undefined}) {
+    const xoaBinhLuan = async () => {
+        const bearerToken = await getBearerToken();
 
-            axios.delete(urlXoaBinhLuan, {headers: { Authorization: bearerToken}})
-                    .then(() => {
-                        layCacBinhLuans();
-                    })
-                    .catch(() => {})
-        })
-        .catch(() => {})
+        if (!bearerToken) {
+            return;
+        }
+
+        const urlXoaBinhLuan = url(`api/binhluan/${binhLuan.bL_Id}`);
+
+        try {
+            await axios.delete(urlXoaBinhLuan, {headers: { Authorization: bearerToken}});
+            layCacBinhLuans();
+        }catch {}
     }
+    
     return (
         <View>
-            {userLogin.id === binhLuan.bL_NguoiTao_Id || userLogin.role === ROLE_ADMIN ? 
+            {userLogin 
+            ? userLogin?.id === binhLuan.bL_NguoiTao_Id || userLogin?.role === ROLE_ADMIN 
+                    ? 
             (<View>
                 <TouchableOpacity onPress={xoaBinhLuan}>
-                    <Text style={{backgroundColor: 'red', width: '8%'}}>Xóa</Text>
+                    <Text style={{backgroundColor: 'red', width: width, borderRadius: 8}}>Xóa</Text>
                 </TouchableOpacity>
             </View>)
             :
-            (<View></View>)}
-            
+            (<View></View>)
+            : (<View></View>)}
         </View>
     )
 }
